+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasetop.py
@@ -100,7 +100,7 @@ class CreateDatasetResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2026-09-07T12:50:42.661Z")
+    updated: Optional[datetime] = parse_datetime("2026-09-08T00:03:29.341Z")
     r"""The date and time the resource was last updated"""
 
     @model_serializer(mode="wrap")