+++ packages/orq-rc/src/orq_ai_sdk/models/getevalop.py
@@ -14,3 +14,11 @@ class GetEvalRequest(BaseModel):
     id: Annotated[
         str, FieldMetadata(path=PathParamMetadata(style="simple", explode=False))
     ]
+
+
+class GetEvalResponseBodyTypedDict(TypedDict):
+    r"""OK"""
+
+
+class GetEvalResponseBody(BaseModel):
+    r"""OK"""