+++ packages/orq-rc/src/orq_ai_sdk/models/updatetoolop.py
@@ -1107,7 +1107,7 @@ class UpdateToolResponseBodyCodeExecutionTool(BaseModel):
     code_tool: UpdateToolResponseBodyCodeTool
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "tool_01M1XYQRHQNN4V9ZMCZBRMQK60"
+        "tool_01M1Z57K8AWE68QY0RBYCYM3S4"
     )
 
     display_name: Optional[str] = None
@@ -1412,7 +1412,7 @@ class UpdateToolResponseBodyHTTPTool(BaseModel):
     http: UpdateToolResponseBodyHTTP
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "tool_01M1XYQRHPNB4AE6ZKSP25MG5A"
+        "tool_01M1Z57K88CDS5SZ9Y50HN3W1A"
     )
 
     display_name: Optional[str] = None
@@ -1606,7 +1606,7 @@ class UpdateToolResponseBodyJSONSchemaTool(BaseModel):
     json_schema: UpdateToolResponseBodyJSONSchema
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "tool_01M1XYQRHPBKSWD0WPXFZVXZE0"
+        "tool_01M1Z57K881KEHHYNNK6E9FS12"
     )
 
     display_name: Optional[str] = None
@@ -1804,7 +1804,7 @@ class UpdateToolResponseBodyFunctionTool(BaseModel):
     function: UpdateToolResponseBodyFunction
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "tool_01M1XYQRHN0QC1BS6CQSRKXPK0"
+        "tool_01M1Z57K87Y5PNPDQ0MZGGF674"
     )
 
     display_name: Optional[str] = None