+++ packages/orq-rc/RELEASES.md
@@ -10729,3 +10729,13 @@ Based on:
 - [python v4.15.0-rc.25] packages/orq-rc
 ### Releases
 - [PyPI v4.15.0-rc.25] https://pypi.org/project/orq-ai-sdk/4.15.0-rc.25 - packages/orq-rc
+
+## 2026-09-08 00:06:02
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.796.4 (2.935.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.15.0-rc.26] packages/orq-rc
+### Releases
+- [PyPI v4.15.0-rc.26] https://pypi.org/project/orq-ai-sdk/4.15.0-rc.26 - packages/orq-rc